refactor: extract BaseAgent ABC and ChatAgent from MainAgent (#140)
* refactor: centralize env vars and magic strings into config/constants.py

Create agents/main_agent/config/constants.py with EnvVars, Defaults, and
Prefixes classes. Update all 13 modules to import from the centralized
constants instead of using inline os.getenv() with hardcoded strings.

This eliminates scattered magic strings and provides a single reference
for all configuration. Zero behavior change — all values are identical.

543/543 tests passing.


* refactor: extract BaseAgent ABC and ChatAgent from MainAgent

Split MainAgent into a three-tier hierarchy:
- BaseAgent (ABC): shared init for model config, tools, session, streaming
- ChatAgent(BaseAgent): Strands Agent creation and text streaming
- MainAgent(ChatAgent): backward-compatible alias (pass-through)

All existing callers continue to import and use MainAgent unchanged.
The _build_filtered_tools() helper is extracted from _create_agent() for
reuse by future agent types (SkillAgent, VoiceAgent).

543/543 tests passing — zero behavior change.


---------
